Open Ctrl-Cut in it's own process; closes #3, also resolves #1

+++ ctrl-cut.py
@@ -1,13 +1,21 @@
 #!/usr/bin/python
 
 import os, sys
-from subprocess import call
+from subprocess import call, Popen
 
-svgfile = sys.argv[-1]
-epsfile = svgfile+".eps"
-null = open(os.devnull, "w")
-call(["inkscape", svgfile, "--export-eps="+epsfile], stderr=null)
-call(["ctrl-cut", epsfile], stderr=null)
-null.close()
-os.remove(svgfile)
-os.remove(epsfile)
+def CtrlCut(epsfile):
+	call(["ctrl-cut", epsfile])
+	os.remove(epsfile)
+
+if __name__ == '__main__':
+	if (len(sys.argv) > 2):
+		if (sys.argv[2] == "child"):
+			CtrlCut(sys.argv[1],)
+	else:
+		svgfile = sys.argv[-1]
+		epsfile = svgfile+".eps"
+		null = open(os.devnull, "w")
+		call(["inkscape", svgfile, "--export-eps="+epsfile], stderr=null)
+		Popen(["python", "ctrl-cut.py", epsfile, "child"], stdout=null, stderr=null)
+		null.close()
+		os.remove(svgfile)
